sys: constify various pointer params
These are all cases where the C API takes a mutable pointer when it
shouldn't - it is not mutating the object and shouldn't be expected to
per the purpose of the functions. Patches for all of these cases have
been sent to PA project. No point in waiting for them to accept those
patches.

Note, there are some functions which may look like they could be
changed, but can't. These include a few which return a mutable pointer
to an inner attribute, for which a seperate immutable version would
be required, and functions which perform an immutable action, but
which also have one or more validation checks which can result in an
immutable function call to store an error value.