Promote `self` to mutable for some methods
Although our objects in these cases can be used immutably (no change
in the wrapped pointer), we should reflect mutability needs of the
operation being performed on the underlying C object.

Note, there are some functions which may look like they could be
immutable, but can't be. These include a few which return a mutable
pointer to an inner attribute, for which a seperate immutable version
would be required in the C API, and functions which perform an immutable
action, but which also have one or more validation checks which can
result in an immutable function call to store an error value.