news: remove dependency note from v1.0.5, confusing
All crates updated there dependencies. The note was intended to make it
clear that this was the only change on three of the crates (done to bring
everything inline and thus avoid users of crates ending up with multiple
versions of some of them), but was probably just confusing. :/
A note like this is not really needed, users can just assume this is the
case when there are no specific change notes yet a new version issued.