context: make self immutable for some methods
Previous commits have:
1) Changed a large number of C API functions to take const pointers,
inline with patches sent in for PulseAudio itself (which have now been
accepted on their 'next' branch), reflecting the fact that these functions
do not change the objects pointed to and were not designed to ever change
them.
2) Changed a lot of immutable `self` references to mutable, where the
underlying function correctly takes a mutable pointer - considering that
really we should reflect the underlying mutability of the C API functions,
not use an immutable self just because our wrapper object isn't being
changed.

The work patching PA itself left a handful of context related functions
unchanged, despite the purpose of those functions having no intention of
modifying their primary object. The reason for this lies with validation
checks, which may modify the error property of the context object. Many
other API functions do this, but most do so via a pointer and thus could
be changed, but these context functions accessed the object directly and
thus could not be changed. It is not ideal that such an implementation
detail leaks to the outside world in forcing mutability in the public API
in this way, and really we could be justified and safe in just ignoring it
and pretending that they are const in our description of the API. However,
I have since created and submitted patches to PA that move the error
property behind a pointer, and then change this handful of holdouts to
const, solving the problem. The basic 'move behind pointer' patch has
already been accepted on PA's 'next' branch, and the rest are to follow,
pending agreement over a minor implementation detail. There is no point
in waiting for a new version of PA to be released, we can go ahead and
constify our description of the API now.