update copy of C includes
to PA v12

Asside from the version update, the changes are to documentation only -
largely addition of return info, typo fixes, and de-duplication of buffer
attribute documentation, from a couple of patches I sent in myself
previously.

No changes are needed in the binding to reflect these changes, they were
already accounted for in v1.0.