sys: constify pa_mainloop_api refs
The PA devs tell me that pa_mainloop_api is supposed to be immutable.
I have sent in a bunch of patches to constify this in PA itself, some of
which require an API bump. Since nobody should be modifying it anyway,
let's go ahead and constify in our API here.