context: take mainloop ref in creation instead of api ref
The primary motivation for this was it being another step towards
removing exposure of the mainloop-api vtable, but it turns out
that that may need to remain for allowing creation of custom
mainloop implementations. However, taking a mainloop ref is
still a simplification for users anyhow, hiding the vtable detail.