operation: fix known possible memory leak
If the user calls a function taking a single-use callback, which returns
an `Operation` object, then calls `cancel` with it, this (deliberately)
prevents the callback from being executed. The problem is that with the
current 'closure' based callback design, if the callback (proxy) is never
executed, nothing frees the closure that the C API is holding as a raw
pointer. This is a known memory leak, and here we fix it.

To solve this problem, we essentially just have the `Operation` object
keep a copy of the pointer for the closure, and if the user tries to
cancel, then it preforms the necessary destruction.

Giving thorough consideration to potential problems:
1) With the synchronous nature of function call and callback execution in
the standard mainloop, and also with the threaded mainloop when `lock()`
is correctly used, no general problems are expected.
2) Users should not call `cancel()` more than once, but what if they do?
With the raw C API, either the ref count assertion may cause a crash,
or otherwise the state already being 'cancelled' will cause it to return
early. with our binding layer on top, we `take()` the pointer out of an
`Option` wrapper, thus preventing freeing more than once.
3) The user should never call `cancel()` within the callback itself. This
would make absolutely no sense, and doing so will break things at the C API
level if you try. We do nothing to explicitly try to block this in the
binding layer.