events: pass event source ref into callback closures
The raw C callbacks receive a pointer to the event source object itself,
but this was not passed on in the new closure callback based design (we
don't want to just pass on a raw pointer after all). We don't know if
users will ever actually want to manage the event source from within
execution of its own callback, but there's no specific reason to take the
option away from users.

The changes made here implement a solution for restoring this feature.
Without it users would be forced to work around it with an ugly mechanism
involving: a wrapper object; an atomic (or Rc); capturing a ref of the
wrapper in the closure; and storing a ref for the event source in the
wrapper after creation. Ugly!

Note, only the proper event source holds saved closures, and no methods
are provided for changing such saved closures, so no problem from that
perspective.