replace include list with exclude list in toml files
It seems that the license file cannot be included from a parent directory
using "../", which was the purpose of using an include list. Instead we
can just use an exclude list to get rid of the files we don't want, and if
the license file is still wanted in the bundled packages (too much bloat?)
sym files could offer the solution.