Add fallback where pkg-config has no *.pc file available
*.pc files may be distributed in *-dev packages, which also include C/C++
header files, and are naturally installed for those working in C/C++ code,
but not so in Rust. Thus we need a fallback for Rust users who do not
happen to have the *-dev package installed (like myself).

Ideally distributions should perhaps change the way *.pc files are packaged
in order to better support Rustaceans.