enable Send+Sync
...for a bunch of types

as requested in https://github.com/jnqnfe/pulse-binding-rust/issues/14

With the threaded mainloop, a lock must be held when using objects; this
is taken by the mainloop dispatcher when executing callbacks, and otherwise
is taken by the user. The user must be careful to take that lock before
using objects within any thread. With that locking mechanism, it should be
safe I presume for these objects to be Send+Sync. It is not ideal that the
user can so easily just forget to grab the lock, but that's the PA design
and it's not certain that we can really do anything to address this.