update to PA v13 changes
note that the 'const pointer' changes in the updated copy of the '.h'
includes were already previously incorporated.

the one change not implemented here is the addition of a threaded-mainloop
method based upon the new pa_threaded_mainloop_once_locked() function,
delayed pending consideration of the safety of it and how to best wrap it.