binding: add capi struct/enum comparison tests
for size and alignment comparison checks (no way to check layout)

reminder: these structs/enums were duplicated between the binding and sys
crates in order to change or add documentation (sys has minimal docs and
binding has full, and in places changes are needed, such as type links).
(type renaming is done also, but that's possible through `use` imports so
was not a basis for this).

ideally we would import the same structs/enums from a shared source file,
with different documentation applied, which would garuantee them being
identical in all ways, but I know of no good means to achieve this.