fix workspace use wrt. version compatibility
having restored the workspace to a proper crate setup in the previous
commit, we now add version compatibility features for control over the
respective features of the crates within it.

(my presumption that features at workspace-compilation level would just be
passed along to the crates within it has proven false).