context: add missing option wrapper to `Context::get_tile_size()` param
The C API function allows a null pointer to be used for this param, while
our binding forced a non-null pointer to be provided.

I noticed an incomplete comment within the function, dating all the way
back to v1.0, intending to explain the reasoning for rejecting allowing
use of NULL through the binding (via an `Option` wrapper on the param).
I must have gotten distracted before either finishing up the comment, or
reviewing whether this stance was really the right one to take.

Reviewing the situation, perhaps I had been confused by the validation
logic within the C function, thinking that it was out of line with its
documentation and did not actually allow a null pointer use on that param?
Either way using the function without a sample spec object is perfectly
legit, so we need to support this by adding the missing `Option` wrapper
on the param.