travis: fix failure
since fixing the workspace to correctly make use of feature flags, things
have been failing with an error regarding a missing symbol, specifically
the one introduced in PA v12 and guarded by a feature flag. the previously
broken use of feature flags with the workspace meant that the symbol never
came into use previously and so no failure.

this simply disables PA v12 support, since it is not yet available in any
travis environment.