travis: properly fix failures
it seems that the `--all` flag may be incompatible with the previous fix
for passing feature flags through to the workspace members. if `--all` is
used then it appears that the default feature flags of the members **are**
used, despite having clearly said for them to not be in the dependencies.
because features are cumulative, they get added, and thus we have PA v12
support when we did not want it :/

let's hope this thus fixes the failures on travis...