travis: bump min version
we now depend upon `glib` and `glib-sys` for the glib-mainloop component,
this requires a newer min version of Rust, which (through experimentation
getting it to compile in travis) happens to be 1.31.

a bump to 1.31 was also pending anyway on the `next` branch to allow for
updating to Rust 2018 `use` conventions, so this accomplishes this and
those changes can be published on master.