Revert "workspace: avoid documenting"
The intention was to exclude the workspace lib itself from being
documented and thus from appearing in the list of documented crates
in generated documentation. It happened to block use of `cargo doc`,
requiring `cargo doc --all` to get any documentation.

As I've just discovered, use of `--all` does not work as intended
wrt. features flags at the workspace level with the new lib-crate
based model (used to actually allow use of features to work as
intended). Thus we must avoid all use of `--all` to get correct
documentation wrt. features and thus we must revert this commit to
actually be able to get documentation. sigh. :/

This reverts commit e217e04932840a9b19c252ed34d7003ddd25a519.