sys: improve build scripts
old logic:
 1) do version-specific check
 2) if failed, make generic fallback link attempt

new logic:
 1) do non-version specific check
 2) certain cases of failure indicate lack of pkg-config, pkg-config being
    disabled, or lack of PA *.pc file; in which case we try the fallback
 3) otherwise (we know we have pkg-config and it knows about PA through
    having the *.pc file) we do a version-specific check and honor the
    result, printing errors to stderr

this means for instance that when run against too old a version of PA
(and both pkg-config is installed and libpulse-dev package is installed
such they the PA *.pc files are available) then we should see an error
describing the version issue in the output, rather than experiencing
missing symbol linker errors if you try to use too new of a feature. i.e.
we actually print the pkg-config error in this design.