add `dox` feature
used for generation of documentation, including the very latest PA support
and bypassing the pkg-config system lib version check

this is useful where wanting to generate and review the latest set of
documentation, which may include details of features from newer versions
of PA than you have installed.

note, there are certain parts of the context::introspection code which
dynamically change depending upon whether PA v5 or PA v6+ support is
selected. with respect to this new `dox` flag we go with showing the
most up to date set of information, i.e. in this case we go with the v6+
view of things.