build: fix linker lib targets for non-pkg-config fallback
under certain conditions, such as lack of pkg-config (use of which we
encourage but do not demand), a simple fallback is used where we try to
just link without it. unfortunately we cannot do so (as I have found on
my Linux based system at least) just by specifying the name "pulse" and
expecting the correct name to be automatically constructed from that (as
the linker attempts to do unless you prefix a `:` to it). this seems to be
perhaps because of the versioning - the correct lib name on Linux is
`libpulse.so.0` not just `libpulse.so`. the *.pc (pkg-config) file seems to
use `pulse0` which suggests that it can actually handle such versioning,
but that was not my experience previously when testing.

here we simply flesh out similar specific target lib names (with a bit of
guess work involved with the help of googling) for Windows and MacOS cases.