travis: bump min version
having to push this quickly due to low battery...

builds are failing on the previous min version 1.31 with errors pointing
to stuff relating to pinning and futures. I believe the `glib` crates used
by the mainloop-glib stuff is responsible. I'm just going to bump the min
here up to 1.39 (same as used by the `glib` crates) for now at least to
resolve the failure...