Print warnings with cargo:warning
`eprintln!` does not appear in Cargo output; however, we can give Cargo
the warning instruction [1], which will display a warning with the specified
message.

[1] https://doc.rust-lang.org/cargo/reference/build-scripts.html#cargowarningmessage

tweaked by committer: removed unnecessary formatting changes that I
disagreed with.

note to self (committer): testing proved that the switch from `eprintln!`
to `println!` was necessary in addition to use of `"cargo:warning=MESSAGE"`
in order to see the output.