binding: improve standard mainloop prepare() timeout type
the underlying C API takes an i32. the param here in the binding was
`<Option<i32>` partly due to not previously knowing the scale, which has
just been clarified as being microseconds in the PA v14 docs, and
specifically `i32` was used rather than `u32` as would have made much more
sense, since `i32` gave a simple way to cap the max value to `i32::MAX`
as it needs to be, and since negative values are perfectly valid for the
C API.

now that we know that the scale is microseconds, it is much cleaner and
clearer to make use of our `MicroSeconds` type here. this requires of
course that we take care to handle values that are too large, which we can
do by simply returning an appropriate error.