sys: add readme files
to prep for addition of badges to be displayed on crates.io pages (and in
github when user is viewing the contents of each crate directory, which
will also automatically display the rendered readme).

besides, the notes of compatibility feature flags may be useful to have
as with the binding layer ones anyway.