readme: link minimum rust version badge to RFC
as done with such badges on crates by the rust core developers.

otherwise, the badge, on gitbuh rendering of the readme files at least,
are automatically linked to the raw file with a very long, messy and
suspicious looking URL, rather than just using the image without any such
link.

the same problem exists for the license badge shown on the main project
readme, but it's not obvious what i could sensibly link that to. it's
not really a big deal, so i guess i'll just leave it.