derive `FromPrimitive` and `ToPrimitive` from `num-derive` on basic enums
thus providing functionality for translating between enum variants and
integers.

this can then be used for instance in the introspection code where we need
to convert data from raw integer form to enum variants, which is currently
being done simply with a transmute which is not ideal since it has the
potential for UB from unexpected values that do not map. we of course
could have implemented better translation before, but this allows it to be
constructed implicitly which is helpful.

we're probably not going to actually use it for most enums, but there's no
harm in making it available broadly.