shrink channels-max constant's type to `u8`
i believe i originally used `usize` simply because i had no idea what type
to assign based simply on the PA header define. however it seems clear to
me now that this relates directly to values for the 'channels' attribute
of the `Spec` (`pa_sample_spec`) type. the type for that is `u8`, and the
value of the constant fits well within that range, so let's change to `u8`.