binding: fix excessively large channel count param types
both `Map` and `ChannelVolumes` hold a fixed size array, exactly the
same as the underlying C types, with the length of that array being
`CHANNELS_MAX`, which is currently set to 32. additionally, they both
include an attribute indicating how much of the arrays are in use,
which in both cases is a `u8`.

there were some functions using a `u32` to take a channel count, which
is silly since the count is limited to the range of a `u8` in those
cases. the use of `u32` simply came from the fact the underlying C
functions they call take an `int`.

there is no need for a type larger than `u8`. changing to `u8` better
reflects to users what sort of range of values are expected.