sample: adjust export of endian-independant format constants
deduplicates the comments; but at the expense of having a third set of
constant declarations.

the real reason for doing this though is as a step towards making the
constants available as associated constants of the `Format` enum. although
there was nothing directly stopping this as it was, you can't assign
deprecation markers to a `pub use` declaration, and putting them on each
individual constant as it was would not only mean twice as many as needed
now but also attributes removing warnings where assigning the values to
the associated constants, unless duplicating the values (bad form).