def: fix broken link and info
overlooked in 10d17f8e699339857daeecb7660251f9e696feb9 and thus broken
since v1.2.

also fixes fact that in the second case, the sentences were repeated twice
for some reason, once with the woord 'sink' and once with the word
'source'. this mistake is present within the very first commit in this
repo, but not in the actual C header. so it seems to be have been some
screw up of mine in the very first version.