ditch `pa_latest` and `pa_latest_common` features, use `pa_v8` as default
the 'latest' aliases were a bad idea, they are a pain to maintain.
 - already i've messed up by forgetting to update `pa_latest` to point to
   `pa_v14` after adding `pa_v14`.
 - additionally it took some time for me to remember that i should update
   `pa_latest_common` to `pa_v13` after `pa_v13` had been around for a
   while.

the whole concept is flawed and not very useful to users. i've decided it's
not worth the trouble and i'm getting rid.

as for the default, i've somewhat arbitrarily decided to choose `pa_v8` as
the new default. this matches the version in the default Ubuntu environment
used for travis builds, which is a couple of big Ubuntu releases behind
now, so PA v8 is pretty solidly "common" now. PA v12 is perhaps ever so
slightly too recent to jump to, even if we did have `pa_v13` as the
default up until now.

support older than v8 is only there really because it was easy to offer.
we don't really _need_ to drop the default interface that far. v8 is a
good place to sit. fundamentally though users have full control over
choice, this is just about how much to put in place, and more importantly
what minimum version to force via pkg-config for users adding a dependency
on use without yet exploring the available version controls. again, v8
seems a good place to sit on.

worth noting at this point is that enhancements are being made to rustdoc
to provide controls such as `#[cfg(doc)]`, allowing detection of compiling
in doc generation mode, which we can use at some point to replace use of
the `dox` feature. additionally there will be means to tag things with the
feature necessary to enable it, with that appearing in the docs. so we
definitely want to move to a position where we always include all versioned
stuff for doc generation.