use `#[cfg(doc)]` for including all feature-guarded stuff in docs
...instead of the `dox` feature flag. the `#[cfg(doc)]` feature was
introduced in Rust v1.41. it should work better in some ways than the
`dox` feature flag.

the `dox` feature flag is now limited to only being used to bypass
the pkg-config check. if we can use `#[cfg(doc)]` for this also then
great, we can ditch `dox`, but i believe i saw a mention somewhere that
it does not work in build scripts. this needs investigating.