mark documentation with feature requirements
the `#[cfg(doc)]` attribute made use of in a previous commit is used as a
means of signalling that items should be enabled when generating
documentation, per the `doc` flag being passed along as a command line
config option. it replaced the existing design of crates like mine making
use of a "dox" feature flag that would only get used under certain
circumstances such as our crate config telling docs.rs to use it. this
replacement however is supplied by the tool that generates the
documentation, so is more reliable.

the new attributes added here are based on the `#[doc(cfg(feature = "foo"))]`
type attributes, which allow tagging items with feature information that
you want included in the generated documentation, to help users understand
that certain things are only available if you enable the specified feature.

however, the implemented attributes are more complex than the example just
given, looking instead like `#[cfg_attr(docsrs, doc(cfg(feature = "foo")))]`.
the reason for this is because the attribute is currently only available
on the nightly version of the compiler. remember, `cfg_attr()` is used to
conditionally apply other attributes. what we are doing here is
conditionally applying these feature attributes only if the config param
"docsrs" is supplied. the docs.rs service make use of the nightly version
of the compiler, and so we can take advantage of that to get the docs.rs
version of the documentation immediately updated with feature tagging,
hence why i'm getting this work done now. in our crate config for docs.rs
we can tell it to make use of a custom config param of "docsrs", thus
enabling these conditional attributes, thus enabling feature tagging.
anyone else with the nightly compiler could also compile with
`--cfg docsrs` passed along and also benefit. remember, config params can
be refered to simply as `#[cfg(foo)]` while for features it is
`#[cfg(feature = "foo")]`, hence why we prefer to use a config param for
this rather than a feature like we did with `dox` previously.
also remember that config params are supplied like `dox` previously.

once the new attribute becomes available on the stable version of Rust, and
it is okay to bump the minimum supported Rust version, we can ditch the
`cfg_attr()` wrapper.

one further thing is required to get this working, and that it to actually
enable the Rust `doc_cfg` feature for the crates, which again we need to do
conditionally since it's nightly only. we do that by adding lines like
this: `#[cfg_attr(docsrs, feature(doc_cfg))]` to the root of each crate.
we will also want to remove this once the feature becomes stable.