replace use of `dox` feature for bypassing build script pkg-config checks
for generation of documentation, the actual system lib is unnecessary. we
do not need to do a pkg-config check for it, hence our `dox` feature
provided a means of bypass.

we already ditched use of that feature for including all feature guarded
crate elements within generated documentation in favour of `#[cfg(doc)]`
in a previous commit, with that capability having shipped in Rust 1.41.

that left the feature only being used in the build scripts. unfortunately
cargo does not currently seem to pass along `--cfg doc` or the equivalent
environment flag (`CARGO_CFG_DOC`) to build scripts, which is problematic.

what has been done here is to use features of the the docs.rs config to
have `--cfg doc` passed along to build scripts for documentation generated
by the docs.rs service. working around the failure of it getting passed
along otherwise.

note, the existing config name `doc` was used for this rather than `docsrs`
as used for `rustdoc-args` since that's designed for only enabling a
nightly only feature for docs.rs, until such time that the feature
stablises (using `doc` for that would cause errors building with the stable
compiler). there is no such problem using `doc` for other things, mirroring
the flag that really should be passed in anyway.

this change allows us to ditch the `dox` feature completely, although only
generation of docs by docs.rs is affected. users wanting to generate docs
locally without having the system lib will now no longer have the `dox`
feature flag there to do that. i doubt anyone's going to be affected by
that though in reality.