readmes: fix mistake
the text added to the sys crates in c47b4fc2636036c51665bff85e0d34c87a95526d
was the copy from the main binding, which was the wrong text for these.
the text is slightly different for the other two binding crates due to them
using multiple crates in the usage instructions, and it should have been
that text which was used for these.