time: fix lack of use of “checked” calculations in certain basic math ops
to thus panic on overflow rather than wrap. in the case of div/rem it's
divide by zero that is important, and i guess this makes no difference to
that.

i believe that i had originally thought that this was already the default
but it seems not. i may have been tricked by the difference between debug
and release builds (i read something suggesting that debug builds do such
checks implicitly, while release builds do not).