ci: add step to output toolchain versions
to double check that the version used in `cargo` commands is the one
selected via the github action-rs action.

as well as the `action-rs/toolchain` action for fetching the tooling, there
is also the `action-rs/cargo` action for executing cargo, and then the
weird `override` param. the override param seems to have something to do
with choosing which copies of the tooling get used, either the system one
or the downloaded and `rustup override` selected one. this is not something
that i am familiar with.

it may be the case that not making use of `action-rs/cargo` that when i
execute `cargo check`, it's using the wrong version of the tooling.
adding this version output should help check.

if there is a problem, then other projects like `rand` are doing it wrong
also, since i learned from them.

note that the `action-rs` people are discussing a v2 which would use
`override` by default...