docs: add missing docs
to try to achieve 100% documentation coverage as reported on docs.rs.

we're very close, having gone from 89% to 99.35% with version 2.20.

there are apparently 7 items remaining.

i note that there calculation of number of items with examples seems very
wrong, stating 3/585, considering all of the examples added to
`MicroSeconds` methods in 2.20.