version: tweak constants
the values should reflect the attributes of the PA version for which the
flag was introduced, not their state in the latest version prior to the
need to introduce the next feature flag.

when i originally set these i was not certain what to do. it is clear now
that them being as they are is problematic for multiple reasons.

note, this is necessary for things to work properly with the change to be
implemented next that introduces version checking functions.