simple: pass `get_latency()` error codes along to caller
don't mask the potentially useful hint as to the cause of the problem.

i picked the `Invalid` variant of `Code` for the hypothetical situation
of the underlying C function indicating success whilst returning a value
of `PA_USEC_INVALID` since it was not obvious otherwise what would be more
appropriate.