ci: add schedule
run a test at the beginning of the first day of every month, such that if
i'm busy working on other things and so haven't done any work on this
project in a while, and a dependency is updated and breaks things, i'll
find out about it automatically at most a month later when a CI failure
email arrives.

the schedule syntax was found at [1]. essentially the cron string holds
five space separated values, for minute; hour; day; month; year, with
asterisks meaning any value can match. (i could switch the schedule to
run every other month by switching the fourth param to `1,3,5,7,9,11`).

[1]: https://docs.github.com/en/actions/reference/events-that-trigger-workflows#scheduled-events