ci: tweak step logic
avoid the 'minimum' conditional logic; having the steps for that outside of
the loop results in the same number of lines, but with greater clarity and
getting to drop the "minimum" non-feature-flag from the flag list.

tweak the comments output - i confused myself with the 'testing for
PulseAudio version compatibility' text.