fix 'unresolved import' issue
it does not work to have these feature guards currently until rustc issue
#8811 ([1]) is fixed since we just get errors running `cargo test` or
`cargo doc` otherwise.

the problem stems from the fact that `--cfg doc` will be used for compiling
the main crate being processed but not its dependencies, thus here the
re-exports are enabled in the binding crate, but not the actual items in
the sys crate, and thus 'unresolved import' errors occur, complaining about
not being able to find the items in the sys crate.

note, it obviously would not hurt to just have these string constants
unguarded, but i felt that it would be perhaps nicer to use the guards and
associated feature documentation to highlight the version requirement for
the use of these two significant properties.

[1]: https://github.com/rust-lang/cargo/issues/8811