add some rustfmt ignore hints
(i'm just playing with rustfmt, not committing to use of it yet, with a
number of bugs filed against the rustfmt project about things i'm not
completely happy with).

this ensures that rustfmt ignores the excessively long single line C function
and type definitions in the sys crates, which i like as they are. i don't
care for rustfmt spreading those out over multiple lines to meet the usual
100 char line length limit.