mainloop-api: privatise attributes of `MainloopInner`
i don't believe that users have any need to mutate these objects after
creation. a creation function and additional helpers have been added to
cover all existing direct access to the attributes in the main and
glib-mainloop crates.

`get_ptr()` has additionally been marked as `unsafe` due to the risks
involved in handing out a copy of the raw pointer, which has a limited
lifetime that users must be careful of to avoid use-after-free.

fixes #42.