time: add `#[repr(transparent)]` to `MicroSeconds`
this does not actually make any difference, since we do not directly
pass the wrapper into any C API functions anywhere, but does no harm
to have. the intention after all is to just get the benefits of the
'new type' pattern, we don't want the memory layout and ABI
implications that result from not having this.