ci: try to fix environment setup failure
...now occurring in scheduled run.

the step where we try to ensure pulseaudio packages are installed is now
failing with 404 errors when trying to download the packages. the package
versions it is trying to download is 13.99.1-1ubuntu3.11. the current
version on the 20.04-updates channel is 13.99.1-1ubuntu3.12. this suggests
that the package has been replaced with a newer one, but the apt lists in
the OS image have not been updated.

perhaps if i were patient i might find it working again in a day or two,
presuming github update the apt lists in their images on a regular schedule
and the failure is just down to a timing issue. hopefully though adding a
step of explicitly running `apt-get update` will fix the problem now and
make this more reliable going forward.