sys: use `ProbeError` from `pkg-config`
pkgconfig 0.3.23 changes added a new error in the case of a probe
failure (see https://github.com/rust-lang/pkg-config-rs/pull/127)

This can cause previously building configuration (ie confs without the
dev package installed for pulseaudio) to fail.

This change uses the new error from pkgconfig instead of the old one.

@jnqnfe: edited to add missing changes to the mainloop-glib crate, and to
cover the new `ProbeFailure` case in addition to the `Failure` case rather
than instead, since without a thorough review I'm not certain whether or
not it is safe to assume that we can actually ignore the `Failure` case
now.